fixed lots of compiler warnings most of them were just formatting issues (spaces instead of tabs) fixed some type cast warnings attempting to follow rule of zero / rule of three/five

+++ engine/src/sysspec.cpp
@@ -111,13 +111,13 @@ static void handle_backtrace()
 				t_symbol = __cxa_demangle(t_info . dli_sname, NULL, NULL, &t_status);
 				if (t_status == 0)
 				{
-					fprintf(stderr, "  in %s @ %u\n", t_symbol, (char *)t_callstack[i] - (char *)t_info . dli_saddr);
+					fprintf(stderr, "  in %s @ %u\n", t_symbol, (unsigned int)((char *)t_callstack[i] - (char *)t_info . dli_saddr));
 					t_handled = true;
 				}
 			}
 			
 			if (!t_handled)
-				fprintf(stderr, "  in %s @ %u\n", t_info . dli_sname, (char *)t_callstack[i] - (char *)t_info . dli_saddr);
+				fprintf(stderr, "  in %s @ %u\n", t_info . dli_sname, (unsigned int)((char *)t_callstack[i] - (char *)t_info . dli_saddr));
 		}
 		else
 			fprintf(stderr, "  in <unknown> @ %p\n", t_callstack[i]);
